requirements.txt
discord.py
motor
requests
dnspython
Pillow
python-dotenv
groq
